bench_app/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
cffi==1.16.0
crispy-bootstrap4==2023.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.2
django-crispy-forms==2.1
django-import-export==3.3.7
et-xmlfile==1.1.0
gunicorn==21.2.0
MarkupPy==1.14
mysqlclient==2.2.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pycparser==2.21
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
xlrd==2.0.1
xlwt==1.3.0
